feat(test): test moduling by functionality
